requirements.txt
adafruit_blinka==8.58.1
adafruit_circuitpython_bitbangio==1.3.18
adafruit_circuitpython_max31855==3.2.24
adafruit_circuitpython_max31856==0.12.3
bottle==0.13.3
gevent==25.5.1
gevent_websocket==0.10.1
paho_mqtt==2.1.0
Requests==2.32.4
websocket_client==1.8.0
